+++ lib/src/alphabet_stick_header_sliverview.dart
@@ -1,4 +1,5 @@
-import 'package:alphabet_index_listview/src/scroll_to_index/scroll_to_index.dart';
+import 'anchor_scroller/anchor_scroll_controller.dart';
+import 'anchor_scroller/anchor_scroll_wrapper.dart';
 import 'alphabet_stick_header_stick.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/gestures.dart';
@@ -10,10 +11,10 @@ import 'dart:async';
 ///group list view controller
 class AlphabetHeaderSliverViewController<T> {
   ///scroll controller
-  final AutoScrollController _scrollController;
+  final AnchorScrollController _scrollController;
 
   ///get scroll controller
-  AutoScrollController get scrollController => _scrollController;
+  AnchorScrollController get scrollController => _scrollController;
 
   ///header scroll to index provider
   AlphabetHeaderScrollToProvider? _headerScrollToProvider;
@@ -21,31 +22,24 @@ class AlphabetHeaderSliverViewController<T> {
   ///header height provider
   AlphabetHeaderHeightProvider? _headerHeightProvider;
 
-  ///scrolling
-  bool _isScrolling = false;
-
   ///create list view controller
   AlphabetHeaderSliverViewController({
-    AutoScrollController? scrollController,
-  }) : _scrollController = scrollController ?? AutoScrollController();
+    AnchorScrollController? anchorController,
+  }) : _scrollController = anchorController ?? AnchorScrollController();
 
   ///scroll to group
   Future scrollToGroup(
     int groupIndex, {
-    Duration? scrollAnimationDuration,
-    AutoScrollPosition? preferPosition,
+    double scrollSpeed = 20,
+    Curve curve = Curves.linear,
   }) async {
     if (_headerScrollToProvider != null) {
       int index = _headerScrollToProvider!(groupIndex);
-      if (_isScrolling == false) {
-        _isScrolling = true;
-        await _scrollController.scrollToIndex(
-          index,
-          duration: scrollAnimationDuration ?? Duration(milliseconds: 450),
-          preferPosition: preferPosition ?? AutoScrollPosition.begin,
-        );
-        _isScrolling = false;
-      }
+      await _scrollController.scrollToIndex(
+        index: index,
+        scrollSpeed: scrollSpeed,
+        curve: curve,
+      );
     }
   }
 
@@ -53,8 +47,8 @@ class AlphabetHeaderSliverViewController<T> {
   Future scrollToChild(
     int groupIndex,
     int childIndex, {
-    Duration? scrollAnimationDuration,
-    AutoScrollPosition? preferPosition,
+    double scrollSpeed = 20,
+    Curve curve = Curves.linear,
   }) async {
     ///childIndex == 0 ,just scroll to group
     if (childIndex == 0) {
@@ -68,9 +62,9 @@ class AlphabetHeaderSliverViewController<T> {
           ? _headerHeightProvider!(groupIndex)
           : 0;
       await _scrollController.scrollToIndex(
-        index,
-        duration: scrollAnimationDuration ?? Duration(microseconds: 450),
-        preferPosition: preferPosition ?? AutoScrollPosition.begin,
+        index: index,
+        scrollSpeed: scrollSpeed,
+        curve: curve,
         deltaOffset: -deltaOffset,
       );
     }
@@ -271,7 +265,7 @@ class _AlphabetHeaderSliverViewState<T>
       },
       child: CustomScrollView(
         key: _scrollKey,
-        controller: widget.controller._scrollController,
+        controller: widget.controller.scrollController,
         physics: widget.physics,
         cacheExtent: widget.cacheExtent,
         dragStartBehavior: widget.dragStartBehavior,
@@ -310,10 +304,10 @@ class _AlphabetHeaderSliverViewState<T>
                   group.dataList[childIndex],
                 );
               }
-              return AutoScrollTag(
+              return AnchorItemWrapper(
                 index: index,
                 key: ValueKey(_uniqueStr + "." + index.toString()),
-                controller: widget.controller._scrollController,
+                controller: widget.controller.scrollController,
                 child: indexItem,
               );
             },
@@ -343,8 +337,8 @@ class _AlphabetHeaderSliverViewState<T>
         ///get item data
         int groupIndex =
             AlphabetIndexTool.getItemIndexFromGroupPos(widget.dataList, s);
-        AutoScrollTagState<AutoScrollTag>? data =
-            widget.controller._scrollController.tagMap[groupIndex];
+        AnchorItemWrapperState? data =
+            widget.controller._scrollController.itemMap[groupIndex];
         RenderBox? itemBox = data?.context.findRenderObject() as RenderBox?;
         Offset? itemTopOffset = itemBox?.localToGlobal(Offset(0.0, 0.0));
 
@@ -387,8 +381,8 @@ class _AlphabetHeaderSliverViewState<T>
         ///get item data
         int groupIndex =
             AlphabetIndexTool.getItemIndexFromGroupPos(widget.dataList, s);
-        AutoScrollTagState<AutoScrollTag>? data =
-            widget.controller._scrollController.tagMap[groupIndex];
+        AnchorItemWrapperState? data =
+            widget.controller._scrollController.itemMap[groupIndex];
         RenderBox? itemBox = data?.context.findRenderObject() as RenderBox?;
         Offset? itemTopOffset = itemBox?.localToGlobal(Offset(0.0, 0.0));